requirements.txt
Flask==3.1.1
pymongo==4.12.1
python-dotenv==1.1.0
bcrypt==4.3.0
PyJWT==2.10.1
gunicorn==23.0.0